Complete category details logic and showing sleep screen

+++ Time/Shared/Screens/Home/CategoryModel.swift
@@ -24,7 +24,7 @@ struct CategoryModel {
     
     var date: Date = Date()
     // category and tracked seconds
-    var data: [String: Int] = [:]
+    var data: [String: Double] = [:]
 }
 
 extension CategoryModel: Codable {}
